# docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  request.UpdatePostRequest:
    properties:
      caption:
        example: Had an amazing trip to the mountains!
        minLength: 1
        type: string
    required:
    - caption
    type: object
...
  response.CreatePostResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/response.Post'
      status:
        example: success
        type: string
    type: object
...
  response.DeletePostData:
    properties:
      message:
        example: Post deleted successfully
        type: string
    type: object
  response.DeletePostResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/response.DeletePostData'
      status:
        example: success
        type: string
    type: object
  response.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
  response.GetAllPostsResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/response.Post'
        type: array
      status:
        example: success
        type: string
    type: object
  response.GetAllUsersResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/response.User'
        type: array
      status:
        example: success
        type: string
    type: object
...
  response.GetPostByIDResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/response.Post'
      status:
        example: success
        type: string
    type: object
  response.GetUserByIDResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/response.User'
      status:
        example: success
        type: string
    type: object
...
  response.Post:
    properties:
      caption:
        example: Had an amazing day at the beach!
        type: string
      created_at:
        example: "2025-01-31T12:00:00Z"
        type: string
      id:
        example: f9d6b52a-76a1-4b2b-9229-4c8db23a5ef2
        type: string
      image_url:
        example: https://example.com/images/beach.jpg
        type: string
      updated_at:
        example: "2025-01-31T12:30:00Z"
        type: string
      user_id:
        example: 2e0850c7-d213-4a91-9b78-bb86e3a6f0d3
        type: string
    type: object
...
  response.SearchPostsResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/response.Post'
        type: array
      status:
        example: success
        type: string
    type: object
  response.SearchUsersResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/response.User'
        type: array
      status:
        example: success
        type: string
    type: object
  response.UpdatePostResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/response.Post'
      status:
        example: success
        type: string
    type: object
  response.UpdateUserResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/response.User'
      status:
        example: success
        type: string
    type: object
  response.User:
    properties:
      bio:
        example: Hi there!
        type: string
      created_at:
        example: "2025-01-31T12:00:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: 3d5a8b92-f1c5-4dbe-a2a7-1d9a8c743e9b
        type: string
      image_url:
        example: https://example.com/profile.jpg
        type: string
      updated_at:
        example: "2025-01-31T12:30:00Z"
        type: string
      username:
        example: john_doe
        type: string
    type: object
host: localhost:8084
info:
  contact: {}
  description: This is a RESTful API for a simple social media application. It allows
    users to manage their posts, including creating, updating, and deleting posts,
    and provides authentication using JWT. The API is built using the Fiber framework
    and interacts with a PostgreSQL database.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FIBER STARTER API
  version: "1.0"
paths:
...
  /posts:
    get:
      description: Get a list of all posts, along with details like the user who created
        them, the caption, image URL, and timestamps.
      produces:
      - application/json
      responses:
        "200":
          description: Successful fetch posts response
          schema:
            $ref: '#/definitions/response.GetAllPostsResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get all posts
      tags:
      - posts
    post:
      consumes:
      - multipart/form-data
      description: Create a new post with a caption. Optionally, you can upload an
        image. If an image is uploaded, its URL will be returned in the response.
        Requires JWT authentication.
      parameters:
      - description: Post caption
        in: formData
        name: caption
        required: true
        type: string
      - description: Post image (optional)
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Successful image upload response
          schema:
            $ref: '#/definitions/response.CreatePostResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new post
      tags:
      - posts
  /posts/{id}:
    delete:
      description: Delete a post by its ID. Only the post creator is allowed to delete
        it. Requires JWT authentication.
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Successful delete post response
          schema:
            $ref: '#/definitions/response.DeletePostResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a post
      tags:
      - posts
    get:
      description: Get a post by its unique ID, including the caption, image URL,
        and timestamps.
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful fetch post response
          schema:
            $ref: '#/definitions/response.GetPostByIDResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get a post by ID
      tags:
      - posts
    patch:
      consumes:
      - application/json
      description: Update only the caption of an existing post. Only the post creator
        is allowed to make this change. Requires JWT authentication.
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body with updated caption
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful update response
          schema:
            $ref: '#/definitions/response.UpdatePostResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing post's caption
      tags:
      - posts
...
  /posts/user/{user_id}:
    get:
      description: Get all posts made by a specific user, including the caption, image
        URL, and timestamps.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful fetch posts by user response
          schema:
            $ref: '#/definitions/response.GetAllPostsResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get all posts by a specific user
      tags:
      - posts
  /search/posts:
    get:
      description: Search for posts that match a given query, such as a keyword in
        the caption or content.
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful search response
          schema:
            items:
              $ref: '#/definitions/response.SearchPostsResponse'
            type: array
        "400":
          description: Invalid query parameter
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Search posts
      tags:
      - search
  /search/users:
    get:
      description: Search for users by their name or email.
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful search response
          schema:
            items:
              $ref: '#/definitions/response.SearchUsersResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Search users
      tags:
      - search
  /users:
    get:
      description: Retrieve a list of all users from the database.
      produces:
      - application/json
      responses:
        "200":
          description: Successful fetch users response
          schema:
            $ref: '#/definitions/response.GetAllUsersResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get all users
      tags:
      - users
    patch:
      consumes:
      - multipart/form-data
      description: Update the bio, image_url, and/or username of the authenticated
        user. All fields are optional. If a field is not provided, the existing value
        will be retained.
      parameters:
      - description: Updated username (optional)
        in: formData
        name: username
        type: string
      - description: Updated bio (optional)
        in: formData
        name: bio
        type: string
      - description: Updated image (optional)
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Successful update user response
          schema:
            $ref: '#/definitions/response.UpdateUserResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized or invalid token
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user information
      tags:
      - users
  /users/{id}:
    get:
      description: Retrieve the details of a specific user by their ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful fetch user by ID response
          schema:
            $ref: '#/definitions/response.GetUserByIDResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get user details by ID
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"